Various tweaks
Expose trade refresh time for merchants so it can be consumed by IWCB.

Add looting modifier to copper ring drop.

Make guns look less odd in first person while using.
