+++ gradle.properties
@@ -2,7 +2,7 @@ org.gradle.jvmargs=-Xmx6G
 org.gradle.daemon=true
 org.gradle.caching=true
 org.gradle.parallel=true
-mod_version=1.20.1-1.24.1
+mod_version=1.20.1-1.24.1-SNAPSHOT-2023.07.03-1
 mod_group_id=tech.anonymoushacker1279.immersiveweapons
 mappings_channel=parchment
 mappings_version=2023.06.26-1.20.1
