+++ src/main/java/tech/anonymoushacker1279/immersiveweapons/entity/npc/SkeletonMerchantEntity.java
@@ -3,30 +3,21 @@ package tech.anonymoushacker1279.immersiveweapons.entity.npc;
 import com.google.common.collect.ImmutableMap;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
-import net.minecraft.server.level.ServerLevel;
 import net.minecraft.stats.Stats;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.entity.*;
-import net.minecraft.world.entity.ai.goal.*;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.npc.AbstractVillager;
 import net.minecraft.world.entity.npc.VillagerTrades;
 import net.minecraft.world.entity.npc.VillagerTrades.ItemListing;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
-import net.minecraft.world.item.trading.MerchantOffer;
-import net.minecraft.world.item.trading.MerchantOffers;
 import net.minecraft.world.level.Level;
-import tech.anonymoushacker1279.immersiveweapons.entity.GrantAdvancementOnDiscovery;
 import tech.anonymoushacker1279.immersiveweapons.entity.npc.trades.ItemsForEmeralds;
 import tech.anonymoushacker1279.immersiveweapons.init.ItemRegistry;
 
-import javax.annotation.Nullable;
-
-public class SkeletonMerchantEntity extends AbstractVillager implements GrantAdvancementOnDiscovery {
-
-	private int timeUntilRefreshTrades = 24000; // One day
+public class SkeletonMerchantEntity extends AbstractMerchantEntity {
 
 	public static final Int2ObjectMap<ItemListing[]> TRADES = new Int2ObjectOpenHashMap<>(ImmutableMap.of(
 			1, new VillagerTrades.ItemListing[]{
@@ -45,17 +36,7 @@ public class SkeletonMerchantEntity extends AbstractVillager implements GrantAdv
 					new ItemsForEmeralds(ItemRegistry.COBALT_PIKE.get(), 10, 1, 3),
 					new ItemsForEmeralds(ItemRegistry.MOLOTOV_COCKTAIL.get(), 6, 2, 4),
 					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE.get(), 3, 2, 6),
-					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE_GREEN.get(), 3, 2, 6),
-					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE_RED.get(), 3, 2, 6),
-					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE_BLUE.get(), 3, 2, 6),
-					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE_PURPLE.get(), 3, 2, 6),
-					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE_YELLOW.get(), 3, 2, 6),
 					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE_ARROW.get(), 3, 2, 6),
-					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE_ARROW_GREEN.get(), 3, 2, 6),
-					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE_ARROW_RED.get(), 3, 2, 6),
-					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE_ARROW_BLUE.get(), 3, 2, 6),
-					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE_ARROW_PURPLE.get(), 3, 2, 6),
-					new ItemsForEmeralds(ItemRegistry.SMOKE_GRENADE_ARROW_YELLOW.get(), 3, 2, 6),
 					new ItemsForEmeralds(Items.RECOVERY_COMPASS, 16, 1, 2)},
 			3, new VillagerTrades.ItemListing[]{
 					new ItemsForEmeralds(ItemRegistry.DEADEYE_PENDANT.get(), 32, 1, 1),
@@ -69,22 +50,6 @@ public class SkeletonMerchantEntity extends AbstractVillager implements GrantAdv
 
 	public SkeletonMerchantEntity(EntityType<? extends AbstractVillager> entityType, Level level) {
 		super(entityType, level);
-	}
-
-	@Override
-	protected void registerGoals() {
-		goalSelector.addGoal(0, new FloatGoal(this));
-		goalSelector.addGoal(1, new TradeWithPlayerGoal(this));
-		goalSelector.addGoal(1, new LookAtTradingPlayerGoal(this));
-		goalSelector.addGoal(4, new MoveTowardsRestrictionGoal(this, 0.35D));
-		goalSelector.addGoal(8, new WaterAvoidingRandomStrollGoal(this, 0.35D));
-		goalSelector.addGoal(9, new InteractGoal(this, Player.class, 3.0F, 1.0F));
-		goalSelector.addGoal(10, new LookAtPlayerGoal(this, Mob.class, 8.0F));
-	}
-
-	@Override
-	public boolean showProgressBar() {
-		return false;
 	}
 
 	@Override
@@ -108,74 +73,7 @@ public class SkeletonMerchantEntity extends AbstractVillager implements GrantAdv
 	}
 
 	@Override
-	public void aiStep() {
-		super.aiStep();
-		checkForDiscovery(this);
-	}
-
-	@Override
-	public void tick() {
-		super.tick();
-
-		// Lower the trade refresh cooldown
-		if (timeUntilRefreshTrades > 0) {
-			timeUntilRefreshTrades--;
-		} else {
-			updateTrades();
-			timeUntilRefreshTrades = 24000;
-		}
-	}
-
-	@Override
-	protected void rewardTradeXp(MerchantOffer offer) {
-		if (offer.shouldRewardExp()) {
-			int xp = 8 + random.nextInt(6);
-			level().addFreshEntity(new ExperienceOrb(level(), getX(), getY() + 0.5D, getZ(), xp));
-		}
-	}
-
-	@Override
-	public boolean removeWhenFarAway(double distanceToClosestPlayer) {
-		return false;
-	}
-
-	@Override
-	public boolean isPersistenceRequired() {
-		return true;
-	}
-
-	@Override
-	protected void updateTrades() {
-		// Clear existing trades
-		getOffers().clear();
-
-		VillagerTrades.ItemListing[] commonItemListings = TRADES.get(1);
-		VillagerTrades.ItemListing[] rareItemListings = TRADES.get(2);
-		VillagerTrades.ItemListing[] epicItemListings = TRADES.get(3);
-
-		if (commonItemListings != null && rareItemListings != null && epicItemListings != null) {
-			MerchantOffers offers = getOffers();
-			addOffersFromItemListings(offers, commonItemListings, 3);
-
-			int i = random.nextInt(rareItemListings.length);
-			VillagerTrades.ItemListing rareItemListing = rareItemListings[i];
-			MerchantOffer rareOffer = rareItemListing.getOffer(this, random);
-			if (rareOffer != null) {
-				offers.add(rareOffer);
-			}
-
-			int i2 = random.nextInt(epicItemListings.length);
-			VillagerTrades.ItemListing epicItemListing = epicItemListings[i2];
-			MerchantOffer epicOffer = epicItemListing.getOffer(this, random);
-			if (epicOffer != null) {
-				offers.add(epicOffer);
-			}
-		}
-	}
-
-	@Nullable
-	@Override
-	public AgeableMob getBreedOffspring(ServerLevel level, AgeableMob otherParent) {
-		return null;
+	public Int2ObjectMap<ItemListing[]> getTrades() {
+		return TRADES;
 	}
 }
