+++ src/main/java/tech/anonymoushacker1279/immersiveweapons/entity/npc/SkygazerEntity.java
@@ -3,16 +3,15 @@ package tech.anonymoushacker1279.immersiveweapons.entity.npc;
 import com.google.common.collect.ImmutableMap;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
-import net.minecraft.server.level.ServerLevel;
 import net.minecraft.stats.Stats;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
-import net.minecraft.world.entity.*;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.attributes.Attributes;
-import net.minecraft.world.entity.ai.goal.*;
 import net.minecraft.world.entity.ai.targeting.TargetingConditions;
 import net.minecraft.world.entity.npc.AbstractVillager;
 import net.minecraft.world.entity.npc.VillagerTrades;
@@ -22,21 +21,15 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
-import net.minecraft.world.item.trading.MerchantOffer;
-import net.minecraft.world.item.trading.MerchantOffers;
 import net.minecraft.world.level.Level;
-import tech.anonymoushacker1279.immersiveweapons.entity.GrantAdvancementOnDiscovery;
 import tech.anonymoushacker1279.immersiveweapons.entity.monster.StarmiteEntity;
 import tech.anonymoushacker1279.immersiveweapons.entity.npc.trades.*;
 import tech.anonymoushacker1279.immersiveweapons.init.*;
 
-import javax.annotation.Nullable;
 import java.util.*;
 import java.util.Map.Entry;
 
-public class SkygazerEntity extends AbstractVillager implements GrantAdvancementOnDiscovery {
-
-	private int timeUntilRefreshTrades = 24000; // One day
+public class SkygazerEntity extends AbstractMerchantEntity {
 
 	public static final Int2ObjectMap<ItemListing[]> TRADES = new Int2ObjectOpenHashMap<>(ImmutableMap.of(
 			1, new VillagerTrades.ItemListing[]{
@@ -70,19 +63,8 @@ public class SkygazerEntity extends AbstractVillager implements GrantAdvancement
 	}
 
 	@Override
-	protected void registerGoals() {
-		goalSelector.addGoal(0, new FloatGoal(this));
-		goalSelector.addGoal(1, new TradeWithPlayerGoal(this));
-		goalSelector.addGoal(1, new LookAtTradingPlayerGoal(this));
-		goalSelector.addGoal(4, new MoveTowardsRestrictionGoal(this, 0.35D));
-		goalSelector.addGoal(8, new WaterAvoidingRandomStrollGoal(this, 0.35D));
-		goalSelector.addGoal(9, new InteractGoal(this, Player.class, 3.0F, 1.0F));
-		goalSelector.addGoal(10, new LookAtPlayerGoal(this, Mob.class, 8.0F));
-	}
-
-	@Override
-	public boolean showProgressBar() {
-		return false;
+	public Int2ObjectMap<ItemListing[]> getTrades() {
+		return TRADES;
 	}
 
 	@Override
@@ -176,72 +158,6 @@ public class SkygazerEntity extends AbstractVillager implements GrantAdvancement
 	}
 
 	@Override
-	public void aiStep() {
-		super.aiStep();
-		checkForDiscovery(this);
-	}
-
-	@Override
-	public void tick() {
-		super.tick();
-
-		// Lower the trade refresh cooldown
-		if (timeUntilRefreshTrades > 0) {
-			timeUntilRefreshTrades--;
-		} else {
-			updateTrades();
-			timeUntilRefreshTrades = 24000;
-		}
-	}
-
-	@Override
-	protected void rewardTradeXp(MerchantOffer offer) {
-		if (offer.shouldRewardExp()) {
-			int xp = 6 + random.nextInt(8);
-			level().addFreshEntity(new ExperienceOrb(level(), getX(), getY() + 0.5D, getZ(), xp));
-		}
-	}
-
-	@Override
-	public boolean removeWhenFarAway(double distanceToClosestPlayer) {
-		return false;
-	}
-
-	@Override
-	public boolean isPersistenceRequired() {
-		return true;
-	}
-
-	@Override
-	protected void updateTrades() {
-		// Clear existing trades
-		getOffers().clear();
-
-		VillagerTrades.ItemListing[] commonItemListings = TRADES.get(1);
-		VillagerTrades.ItemListing[] rareItemListings = TRADES.get(2);
-		VillagerTrades.ItemListing[] epicItemListings = TRADES.get(3);
-
-		if (commonItemListings != null && rareItemListings != null && epicItemListings != null) {
-			MerchantOffers offers = getOffers();
-			addOffersFromItemListings(offers, commonItemListings, 3);
-
-			int i = random.nextInt(rareItemListings.length);
-			VillagerTrades.ItemListing rareItemListing = rareItemListings[i];
-			MerchantOffer rareOffer = rareItemListing.getOffer(this, random);
-			if (rareOffer != null) {
-				offers.add(rareOffer);
-			}
-
-			int i2 = random.nextInt(epicItemListings.length);
-			VillagerTrades.ItemListing epicItemListing = epicItemListings[i2];
-			MerchantOffer epicOffer = epicItemListing.getOffer(this, random);
-			if (epicOffer != null) {
-				offers.add(epicOffer);
-			}
-		}
-	}
-
-	@Override
 	public boolean hurt(DamageSource source, float amount) {
 		// Give the attacker blindness and summon 2-3 Starmite entities on the attacker
 		if (source.getDirectEntity() instanceof LivingEntity attacker) {
@@ -262,11 +178,4 @@ public class SkygazerEntity extends AbstractVillager implements GrantAdvancement
 
 		return super.hurt(source, amount);
 	}
-
-	@Nullable
-	@Override
-	public AgeableMob getBreedOffspring(ServerLevel level, AgeableMob otherParent) {
-		return null;
-	}
-
 }
