+++ src/main/resources/data/immersiveweapons/loot_modifiers/copper_ring.json
@@ -5,5 +5,6 @@
 		}
 	],
 	"drop_chance": 0.05,
+	"looting_multiplier": 0.02,
 	"type": "immersiveweapons:copper_ring"
 }
